resources/test/tests/promise-with-sync.html

<!DOCTYPE HTML>
<html>
<head>
<title>Promise Tests and Synchronous Tests</title>
</head>
<body>
<h1>Promise Tests</h1>
<p>This test demonstrates the use of <tt>promise_test</tt> alongside synchronous tests.</p>
<div id="log"></div>
<script src="../../testharness.js"></script>
<script src="../../testharnessreport.js"></script>
<script>
var tests = [];

test(function(t) {
  assert_array_equals(tests, []);
  tests.push(1);
}, 'first synchronous test');

promise_test(function() {
  assert_array_equals(tests, [1]);

  return Promise.resolve()
    .then(function() {
        assert_array_equals(tests, [1]);
        tests.push(2);
      });
}, 'first promise_test');;

test(function(t) {
  assert_array_equals(tests, [1, 2]);
  tests.push(3);
}, 'second synchronous test');

promise_test(function() {
  assert_array_equals(tests, [1, 2, 3]);

  return Promise.resolve()
    .then(function() {
        assert_array_equals(tests, [1, 2, 3]);
      });
}, 'second promise_test');;
</script>
<script type="text/json" id="expected">
{
  "summarized_status": {
    "message": null, 
    "status_string": "OK", 
    "stack": null
  },
  "summarized_tests": [
    {
      "message": null, 
      "properties": {}, 
      "stack": null, 
      "name": "first promise_test", 
      "status_string": "PASS"
    }, 
    {
      "message": null, 
      "properties": {}, 
      "stack": null, 
      "name": "first synchronous test", 
      "status_string": "PASS"
    }, 
    {
      "message": null, 
      "properties": {}, 
      "stack": null, 
      "name": "second promise_test", 
      "status_string": "PASS"
    }, 
    {
      "message": null, 
      "properties": {}, 
      "stack": null, 
      "name": "second synchronous test", 
      "status_string": "PASS"
    }
  ], 
  "type": "complete"
}
</script>
